# api-template/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/courseatclass/add": {
            "post": {
                "description": "add new relationship between course and class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course At Class"
                ],
                "summary": "add new relationship between course and class",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of course",
                        "name": "course_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of class",
                        "name": "class_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.CourseAtClass"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/courseatclass/all": {
            "get": {
                "description": "get all record in table courseatclass",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course At Class"
                ],
                "summary": "get all record in table courseatclass",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.CourseAtClass"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/courseatclass/id": {
            "delete": {
                "description": "delete relationship between course and class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course At Class"
                ],
                "summary": "delete relationship between course and class",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of row in table",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.CourseAtClass"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/courseatclass/{id}": {
            "get": {
                "description": "find CourseAtClass record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course At Class"
                ],
                "summary": "find CourseAtClass record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of row in CourseAtClass",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.CourseAtClass"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CourseAtClass": {
            "type": "object",
            "properties": {
                "classId": {
                    "type": "string"
                },
                "courseId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}